--- a/rpcs3/Emu/Cell/SPUASMJITRecompiler.h
+++ b/rpcs3/Emu/Cell/SPUASMJITRecompiler.h
@@ -19,18 +19,26 @@ private:
 	// ASMJIT runtime
 	::jit_runtime m_asmrt;
 
+	u32 m_base;
+
 	// emitter:
 	asmjit::X86Assembler* c;
 
 	// arguments:
 	const asmjit::X86Gp* cpu;
 	const asmjit::X86Gp* ls;
+	const asmjit::X86Gp* rip;
+	const asmjit::X86Gp* pc0;
+
+	// Native args or temp variables:
+	const asmjit::X86Gp* arg0;
+	const asmjit::X86Gp* arg1;
 	const asmjit::X86Gp* qw0;
 	const asmjit::X86Gp* qw1;
 
 	// temporary:
 	const asmjit::X86Gp* addr;
-	std::array<const asmjit::X86Xmm*, 6> vec;
+	std::array<const asmjit::X86Xmm*, 16> vec;
 
 	// workload for the end of function:
 	std::vector<std::function<void()>> after;
@@ -81,12 +89,11 @@ private:
 	asmjit::X86Mem XmmConst(__m128 data);
 	asmjit::X86Mem XmmConst(__m128i data);
 
+	asmjit::X86Mem get_pc(u32 addr);
 	void branch_fixed(u32 target);
 	void branch_indirect(spu_opcode_t op, bool jt = false, bool ret = true);
 	void branch_set_link(u32 target);
 	void fall(spu_opcode_t op);
-	void save_rcx();
-	void load_rcx();
 
 	void get_events();