… rpcs3/Emu/Cell/SPUThread.cpp
… rpcs3/Emu/Cell/lv2/sys_spu.cpp
… rpcs3/Emu/System.h