Settings:
* Block size: Safe or Mega
* SPU LLVM