@isJuhn 2019-05-10
Spider-Man Shattered Dimensions [BLES00946] stops compiling spu code just after boot, rpcs3 stops responding if you mess with ui a bit. on second boot spu compilation stops on 3/100 spu programs

[RPCS3.log](https://github.com/RPCS3/rpcs3/files/3166730/RPCS3.log)
[spu.log](https://github.com/RPCS3/rpcs3/files/3166735/spu.log)