@legend800 2019-05-10
Nothing but fatals or freezes during SPU compile on my games.