@isJuhn 2019-05-11
Spider-Man Shattered Dimensions [BLES00946] now boots and see a huge improvement in terms of spu compilation. nearly all spu programs are reused now. in master 30 seconds of gameplay used to generate 11k spu programs, with 2k unique hashes. now it's 2113 programs and 2054 unique hashes from the same playtime.