@ruipin 2019-05-11
RDR now generates 7k SPU programs in the same playtime it used to generate about 100k. Stutter is virtually gone when using SPU LLVM.

TLOU has regressed (might be related to the previous LLVM PR though), and no longer gets past the loading screen. UC1 also no longer manages to get to menu. If I do not use Safe, all these games I mentioned seem to crash immediately at startup.