@MsDarkLow 2019-05-11
Fist of the North Star: Ken's Rage - DEMO [NPEB90285]
SPU LLVM - Safe
Got a massive improvement, from basically Infinite SPU cache down to 5k.
There is a lot less stutter. 
Doing one of the "most intensive attacks" it went from an avg of 11fps to 25-30FPS! No longer stuttering whenever an action happens.