@FultonBoss 2019-05-11
Metal Gear Rising Revenganvce got massive improvement with SPU LLVM safe, went from 100k+ SPU cache  to 7k
Also better framerate but still suffers from FIFO desync event eventually making the game hang