@legend800 2019-05-11
Splatterhouse is finally playable on the SPU LLVM! No more constant cache building! Nice work. I think this PR fixes this then: https://github.com/RPCS3/rpcs3/issues/5843

![2019-05-10 19_31_30-Window](https://user-images.githubusercontent.com/20983844/57564058-da67d700-735a-11e9-9b02-6d1aaa1fde5b.png)

Unfortunately, Splatterhouse still gets PPU mem errors after a few mins I found out, just a lot less than ASMJIT. This is being tracked separately here: https://github.com/RPCS3/rpcs3/issues/5837

![2019-05-10 19_31_55-](https://user-images.githubusercontent.com/20983844/57564059-db006d80-735a-11e9-9489-37167543525e.png)

[RPCS3.log](https://github.com/RPCS3/rpcs3/files/3168541/RPCS3.log)