@MSuih 2019-05-11
For some reason I can only launch RDR with empty SPU cache on this PR. Attempting to launch with SPU cache will just cause the game to exit and RPCS3 will immidiately close afterwards, but if cache is cleared first then it will boot and go ingame normally.

[RPCS3.log](https://github.com/RPCS3/rpcs3/files/3168561/RPCS3.log)