@arabek 2019-05-11
Tried out RDR and couldn reproduce it. Works from cold cache and after emulator restart with a primed cache of 3737 spu units compiled. LLVM block size Mega.