@MSuih 2019-05-11
Now that I tested this further I think it's tied to size of SPU cache. I managed to boot at 3k cache but then I played a bit further and on next launch the immidiate exit happened again. Tested this a couple of times and got similar results each time

Mega crashes the game for me, but that happens on master as well.