@legend800 2019-05-11
Updated my post above with a quick pass through some titles, noting some that are broken now on this PR. I'll continue to update it with new commit pushes.