@MarioSonic2987 2019-05-11
Settings used: SPU LLVM in Safe mode.

| Game  | SPU Cache Before | SPU Cache After | Screenshot reference |
| ------------- | ------------- | ------------- | ------------- |
| Grand Theft Auto IV | 15k | 6k | [Screenshot](https://user-images.githubusercontent.com/20027681/57565917-1e43f600-73c6-11e9-8834-0598a43920c9.png) |
| Grand Theft Auto V | 31.5k (loading screen and still compiling) | 8.5k (ingame) | [Screenshot](https://user-images.githubusercontent.com/20027681/57565769-ab398000-73c3-11e9-92a5-a158eece1e76.png) |
| Midnight Club: Los Angeles - Complete Edition | ~23.5k | 4.5k | [Screenshot](https://user-images.githubusercontent.com/20027681/57565974-16388600-73c7-11e9-88d7-6eac0753f286.png) |
| One Piece: Pirate Warriors | ~38k (only in title screen and still compiling) | 5.5k (It no longer stutters when you attack or move and performance goes up to ~50 FPS in cutscenes and ~60 FPS in the first level.) | [Screenshot](https://user-images.githubusercontent.com/20027681/57566184-e8087580-73c9-11e9-8e63-00bb17d7516d.png) |
| Saints Row: The Third - The Full Package | 13k-14k (still compiling) | ~6k | [Screenshot](https://user-images.githubusercontent.com/20027681/57565932-767af800-73c6-11e9-9041-13caf35820d7.png) |

This post will be updated when games are found improved by these changes.