@vsub 2019-05-11
RPCS3 auto closes when closing the game window:
Ar NoSurge
Atelier Shallie
Atelier Escha & Loggy
Ar Tonelico Quga
Persona 5

Eternal Sonata  - the game window auto closes after loading the shaders
Tales of Xillia 1\2 - Fatal Error while using ASMJIT(maybe this is what you mean by "ASMJIT don't work") and rpcs3 auto closes when closing the game window while using LLVM(maybe all games are affected by this)