@Nekotekina 2019-05-11
@vsub Try new commit. ASMJIT is currently just unsupported — it needs more work.