@MSuih 2019-05-11
Whatever that RDR safe mode issue was, it's now fixed with latest commit and I can load the game even with existing SPU cache. Mega mode still has some issues but that's not caused by this PR.