@vsub 2019-05-11
@Nekotekina Yes,all problems mentioned above are fixed and Eternal Sonata spu cache seems to be WAY better.
On master if I load a save at the first town and then the first possible save location and play the tutorial,the spu cache becomes above 100k but with this pr,it is less than 2k

And also the game runs a little fasted on this pr