@Xcedf 2019-05-11
RDR started new game, did some random moves 7,5k On Safe never had it on safe before, but on Mega begining of the game was about 35-40k, while Mega cache from whole game is ~59k
Metal Gear Revengeance Safe - 4,4k, was 55k on Mega just in the begining
Now for the bad things Mega hows fatal error window on every game with this PR
GoW 3, TloU, Infamous are stopped working on this PR