@Xcedf 2019-05-11
Yet i think Mega mode regression is a part of previous Giga PR