@Parthorisian 2019-05-11
WET(2009) - crash on load(before going to menu) [BLUS30403](https://i.imgur.com/nrqO8Qr.png) log
[WetRPCS3.zip](https://github.com/RPCS3/rpcs3/files/3169218/WetRPCS3.zip)