@legend800 2019-05-11
On the latest build, closing emu window no longer crashes emu (on Master). 

Several broken games on Master with SPU LLVM work again (Ridge Racer 7, HOTD4, Razing, Xmen, etc). 

However, Wipeout and Motorstorm Apocolypse are still unstable (even with cleared cache). 

Wipeout black screen upon entergameplay:
[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/3169228/RPCS3.log.gz)

MS Apoc freezes loading menus:
[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/3169230/RPCS3.log.gz)