@Ghostotakoi 2019-05-11
This might fix j-star victory versus