@Xcedf 2019-05-11
With recent fix of Mega mode
did more tests
Heavy Rain:
 Safe 4.5k   Mega 1.6k
RDR:
 Safe: 7.2k  Mega 3.7k
Metal Gear Revengeance:
 Safe: 4.4k   Mega 2.2k
Far Cry 2 anoter game with countless SPU cache
 Safe: 4.2k   Mega 1.9k
Some sane numbers after all