@Xcedf 2019-05-12
GoW 3, Uncharted games Infamous TLoU, it seems they are broken on safe mode only and working on Mega in this PR
I successfully ran Infamous demo on mega on this PR and on master i has ~55-60k on Mega just to go ingame on this PR it's 2,2k