@Xcedf 2019-05-12
Found anothe regression in this PR, in GoW3 on Mega level emu crashes after "Sony presents" intro before fmv intro, tested on Master before just to make sure this isn't anoter Giga PR regression and it works fine
![54](https://user-images.githubusercontent.com/7060586/57576849-97992280-7472-11e9-8025-f4832feffd08.jpg)