@legend800 2019-05-12
GOW3 and GOWA both just black screen during boot screens with Safe, no cache. 

[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/3169794/RPCS3.log.gz)