@raveskirza 2019-05-12
Beyond Two Souls (Mega):

Master:
Menu - 12k SPU cache
Intro   - 28K 

PR:
Menu - 1.6K
Intro   -  5k